[APB-11535] yet another way to address jenkins problems
